Moved update.js into modules and modified test function
